==> notebooks/02_validate_and_save.ipynb <==
import pandas as pd
from pathlib import Path

month = 1
year = 2024

path = Path("..") / "data" / "raw" / f"rides_{year}_{month:02}.csv"

rides = pd.read_csv(path)
rides.head()

rides.info()

rides_cp = rides.copy()

rides_cp.isnull().sum()

filter_not_null = rides_cp.notnull().all(axis=1)

rides_cp.dropna(inplace=True)

rides_cp["started_at"] = pd.to_datetime(rides_cp["started_at"], errors='coerce')
rides_cp["ended_at"] = pd.to_datetime(rides_cp["ended_at"], errors='coerce')

# Now safely subtract datetimes
rides_cp["duration"] = rides_cp["ended_at"] - rides_cp["started_at"]

rides_cp.head()

# Here timedelta is an object in datetime module which makes it easy to perform arthimetic operations on the timestamps including days, hours, 
# minutes etc

print(rides_cp["duration"].quantile(0))

print(rides_cp["duration"].quantile(0.01))

print(rides_cp["duration"].quantile(0.99))

print(rides_cp["duration"].quantile(0.999))


rides_cp["duration"].describe()

duration_filter = (rides_cp["duration"] > pd.Timedelta(0) ) & (rides_cp["duration"] <=pd.Timedelta(hours = 5) )

sum(~duration_filter) # This will tell how many instances we are gonna remove

sorted_df = rides_cp.sort_values(by="started_at", ascending=True)  

# Get the top 10 (smallest) and bottom 10 (largest) values  
top_10 = sorted_df.head(10)  
bottom_10 = sorted_df.tail(10)  

top_10


bottom_10

filter_date_range = (rides_cp["started_at"]>= "2024-01-01") & (rides_cp["started_at"] < "2024-02-01")
sum(~filter_date_range)

final_filter = duration_filter & filter_date_range & filter_not_null

numbers_dropped = final_filter.shape[0] - sum(final_filter) # numbers dropped
numbers_dropped
numbers_dropped/final_filter.shape[0] * 100

rides = rides[final_filter]

rides = rides[["started_at", "start_station_id"]]

rides.rename(columns={
    "started_at": "pickup_datetime",
    "start_station_id": "pickup_location_id"
}, inplace=True)

rides.head()

year = 2024

month = 1

path = Path("..") / "data" / "processed" / f"rides_{year}_{month :02}.csv"

rides.to_csv(path,index = False)
