==> notebooks/10_lgm_with_featureEngineering.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import pandas as pd
from src.config import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
df

from datetime import datetime

from src.data_utils import split_time_series_data

X_train, y_train, X_test, y_test = split_time_series_data(
    df,
    cutoff_date=datetime(2024, 3, 15, 0, 0, 0),
    target_column="target"
)

print(X_train.shape)
print(y_train.shape)
print(X_test.shape)
print(y_test.shape)

past_ride_columns = [c for c in X_train.columns if c.startswith("rides_")]
X_train_only_numeric = X_train[past_ride_columns]
X_test_only_numeric = X_test[past_ride_columns]

## Top 10 Feature Selection Using Feature Importance

import lightgbm as lgb
model = lgb.LGBMRegressor()
model.fit(X_train_only_numeric, y_train)

feature_importances = pd.Series(model.feature_importances_, index=X_train_only_numeric.columns)
top_features = feature_importances.sort_values(ascending=False).head(10).index


X_top = X_train_only_numeric[top_features]

X_top

model_top = lgb.LGBMRegressor()
model_top.fit(X_top, y_train)

from sklearn.metrics import mean_absolute_error
predictions = model_top.predict(X_test_only_numeric[top_features])

test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae:.4f}")

from src.experiments_utils import set_mlflow_tracking, log_model_to_mlflow
from dotenv import load_dotenv
import os
load_dotenv() 

mlflow = set_mlflow_tracking()
log_model_to_mlflow(model, X_test_only_numeric, "LGBMRegressorCitiBike_With_Top10_Features", "mean_absolute_error", score=test_mae)

## Principle Componenet Analysis

from sklearn.decomposition import PCA

pca = PCA(n_components=10)
X_pca = pca.fit_transform(X_train_only_numeric)

# Train model on PCA features
model_pca = lgb.LGBMRegressor()
model_pca.fit(X_pca, y_train)


from sklearn.metrics import mean_absolute_error
predictions = model_pca.predict(pca.transform(X_test_only_numeric))

test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae:.4f}")

from src.experiments_utils import set_mlflow_tracking, log_model_to_mlflow
from dotenv import load_dotenv
import os
load_dotenv() 

mlflow = set_mlflow_tracking()
log_model_to_mlflow(model, X_test_only_numeric, "LGBMRegressorCitiBike_With_PCA10_Features", "mean_absolute_error", score=test_mae)
